+++ recommendations.py
@@ -2,6 +2,7 @@ import json
 import argparse
 import sys
 import getSearchResults
+import recommendationUtils
 
 category = json.load( open( "dumps/category.json", "r" ) )
 
@@ -18,10 +19,6 @@ types = args.type
 if(types !="parent" and  types !="all" and types !="child" ):
   print "please enter correct type which can be child/parent/all"
   sys.exit()
-
-
-# print recTerm
-
 
 if recTerm == None:
   print("please enter a term to get recommendation")
@@ -40,81 +37,22 @@ dict_add = lambda x, y={}: dict_add(x[:-1], y).setdefault(x[-1], {}) if(x) else 
 baseDict = {}
 map(lambda x: dict_add(x, baseDict), [path['location'].split("/") for path in category])
 
-def getMaxDepth(l):
-  maxDepth = 1
-  for data in l:
-    maxDepth = max(maxDepth,len(data.split("/")))
-  return maxDepth
-
-def getMaxDepthofDict(d, level=1):
-    if not isinstance(d, dict) or not d:
-        return level
-    return max(getMaxDepthofDict(d[k], level + 1) for k in d)
-  
-def getPath(element, JSON, path, allPaths,childPath):    
-  if element in JSON:
-    path =  path + element #+ str(JSON[element])
-    # global childPath
-    childPath.append(JSON[element])
-    # print(childPath)
-    # print path
-    allPaths.append(path)
-  for key in JSON:
-    if isinstance(JSON[key], dict):
-      getPath(element, JSON[key],path + key + '/',allPaths,childPath)
-
-def getParentWeightedList(l):
-  maxDepth = getMaxDepth(l)
-  parentWeights =  []
-  for i in range(0,maxDepth):
-    parentWeights.append([])
-  for data in l:
-    i=0
-    for  path in data.split("/"):
-      parentWeights[i].append(path)
-      i=i+1
-  return parentWeights
-
-def getChildWeightedList(childPath):
-  try:
-    maxDepth = getMaxDepthofDict(childPath)
-    # print maxDepth
-    childWeights =  []
-    for i in range(0,maxDepth):
-      childWeights.append([])
-    
-    for r, s in childPath.items():
-      i=0
-      q = []
-      p = r
-      while True:
-          for k, v in s.items():
-              # print('(%s,%s) ' % ('ROOT' if p == r else p, k))
-              # if k != "test" and k != "src":
-              #     print k.replace('_',' ')
-              childWeights[i].append(k)
-              if v:
-                  q.append((k, v))
-          if not q:
-                  break
-          p, s = q.pop(0)
-    return childWeights
-  except:
-    return []
 
 
 allPaths = []
 childPath=[]
-getPath(recTerm,baseDict,'',allPaths,childPath)
+recommendationUtils.getPath(recTerm,baseDict,'',allPaths,childPath)
 if len(childPath)>0:
   childPath = childPath[0]
 # print(allPaths)
 # print(getMaxDepth(allPaths))
 recDict = []
 
+print "childPath"
+print childPath
 if types != "child":
   print "parent"
-  parentWeights = getParentWeightedList(allPaths)
+  parentWeights = recommendationUtils.getParentWeightedList(allPaths)
   maxParentWeight = len(parentWeights)
   for parents in parentWeights:
     for parent in parents:
@@ -123,7 +61,7 @@ if types != "child":
 
 # print childPath
 if types != "parent":  
-  childWeights = getChildWeightedList(childPath)
+  childWeights = recommendationUtils.getChildWeightedList(childPath)
   maxChildWeight = len(childWeights) + 1
 
   for childs in childWeights:
